Validation bill and card

+++ src/main/java/bo/ucb/edu/ingsoft/dto/CardRequest.java
@@ -7,7 +7,7 @@ import java.util.Date;
 public class CardRequest {
 
     private String cardName;
-    private Integer cardNumber;
+    private Long cardNumber;
     private Integer expirationYear;
     private Integer expirationMonth;
     private Integer cvc;
@@ -25,11 +25,11 @@ public class CardRequest {
         this.cardName = cardName;
     }
 
-    public Integer getCardNumber() {
+    public Long getCardNumber() {
         return cardNumber;
     }
 
-    public void setCardNumber(Integer cardNumber) {
+    public void setCardNumber(Long cardNumber) {
         this.cardNumber = cardNumber;
     }
